@@ -21,8 +21,8 @@
   "homepage": "https://github.com/divmain/interlock",
   "dependencies": {
     "ast-types": "^0.7.8",
-    "babel": "^5.6.14",
-    "babel-core": "^5.6.17",
+    "babel": "^5.8.21",
+    "babel-core": "^5.8.22",
     "bluebird": "^2.9.34",
     "chokidar": "^1.0.3",
     "escodegen": "^1.6.1",
@@ -34,11 +34,12 @@
     "source-map": "^0.4.2"
   },
   "devDependencies": {
-    "babel-eslint": "^3.1.26",
+    "babel-eslint": "^4.0.5",
     "chai": "^2.3.0",
     "comment-parser": "^0.3.0",
-    "eslint": "^0.24.1",
-    "eslint-config-defaults": "^3.1.0",
+    "eslint": "^1.1.0",
+    "eslint-config-defaults": "^4.2.0",
+    "eslint-plugin-filenames": "^0.1.2",
     "esquery": "^0.4.0",
     "estraverse": "^4.1.0",
     "glob": "^5.0.14",
