@@ -1,10 +1,10 @@
 /* eslint-disable max-nested-callbacks,no-new */
 
-var path = require("path");
+import path from "path";
 
-var resolve = require("../../lib/resolve");
+import resolve from "../../lib/resolve";
 
-var baseDir = path.resolve(__dirname, "../..");
+const baseDir = path.resolve(__dirname, "../..");
 
 
 describe("lib/resolve", function () {
@@ -19,35 +19,35 @@ describe("lib/resolve", function () {
   }
 
   it("resolves file in same directory", function () {
-    var resolved = attemptResolve("./index");
+    const resolved = attemptResolve("./index");
     expect(resolved).to.have.property("ns", "interlock");
     expect(resolved).to.have.property("nsPath", "lib/index.js");
     expect(resolved.path).to.equal(path.join(baseDir, "lib/index.js"));
   });
 
   it("resolves file in sub-directory", function () {
-    var resolved = attemptResolve("./compile/construct/index");
+    const resolved = attemptResolve("./compile/construct/index");
     expect(resolved).to.have.property("ns", "interlock");
     expect(resolved).to.have.property("nsPath", "lib/compile/construct/index.js");
     expect(resolved.path).to.equal(path.join(baseDir, "lib/compile/construct/index.js"));
   });
 
   it("resolves current directory", function () {
-    var resolved = attemptResolve("./");
+    const resolved = attemptResolve("./");
     expect(resolved).to.have.property("ns", "interlock");
     expect(resolved).to.have.property("nsPath", "lib/index.js");
     expect(resolved.path).to.equal(path.join(baseDir, "lib/index.js"));
   });
 
   it("resolves a file in a separate directory branch", function () {
-    var resolved = attemptResolve("../example/build");
+    const resolved = attemptResolve("../example/build");
     expect(resolved).to.have.property("ns", "interlock");
     expect(resolved).to.have.property("nsPath", "example/build.js");
     expect(resolved.path).to.equal(path.join(baseDir, "example/build.js"));
   });
 
   it("resolves a node_modules package", function () {
-    var resolved = attemptResolve("lodash");
+    const resolved = attemptResolve("lodash");
     expect(resolved).to.have.property("ns", "lodash");
     expect(resolved).to.have.property("nsPath", "index.js");
     expect(resolved.path)
@@ -55,7 +55,7 @@ describe("lib/resolve", function () {
   });
 
   it("resolves a file in node_modules package", function () {
-    var resolved = attemptResolve("ast-types/lib/scope.js");
+    const resolved = attemptResolve("ast-types/lib/scope.js");
     expect(resolved).to.have.property("ns", "ast-types");
     expect(resolved).to.have.property("nsPath", "lib/scope.js");
     expect(resolved.path)
