@@ -22,8 +22,9 @@ function getArrayChildren (node, parents) {
   /* eslint-enable block-scoped-var */
 }
 
-export function getTraversableAst (node, key=null, parents=[]) {
-  let type, children;
+export function getTraversableAst (node, key = null, parents = []) {
+  let type;
+  let children;
   if (_.isObject(node) && node.type) {
     type = NODE;
     children = getNodeChildren(node, parents);
