@@ -11,7 +11,7 @@ const bootstrapBundle = Pluggable.promise(function bootstrapBundle (srcPath, bun
     .then(asset => this.loadAst(asset))
     .then(module => {
       return {
-        module: module,
+        module,
         dest: bundlDef.dest,
         entry: isEntryPt,
         includeRuntime: isEntryPt && !bundlDef.excludeRuntime
