@@ -7,7 +7,7 @@ function dedupeExplicit (bundlesWithDupes, modulesByAbsPath) {
   return bundlesWithDupes
     // Generate flat, naive dependency arrays.
     .map(bundle => {
-      let entryModule = modulesByAbsPath[bundle.module.path];
+      const entryModule = modulesByAbsPath[bundle.module.path];
       return _.extend({}, bundle, {
         moduleHashes: [entryModule.hash, ...entryModule.deepDependencies.map((dep) => dep.hash)],
         entryModuleHash: entryModule.hash
