@@ -3,20 +3,20 @@ import _ from "lodash";
 
 import * as Pluggable from "../../pluggable";
 
-function* genBundlesWithImplicit (bundles, implicitBundleDest) {
+function* genBundlesWithImplicit (bundles, implicitBundleDest) { // eslint-disable-line max-len,max-statements
   bundles = bundles.slice();
 
   for (let a = 0; a < bundles.length; a++) {
-    let bundleA = bundles[a];
-    let bundleLengthAtIteration = bundles.length;
+    const bundleA = bundles[a];
+    const bundleLengthAtIteration = bundles.length;
 
     for (let b = a + 1; b < bundleLengthAtIteration; b++) {
-      let bundleB = bundles[b];
-      let intersection = _.intersection(bundleA.moduleHashes, bundleB.moduleHashes);
+      const bundleB = bundles[b];
+      const intersection = _.intersection(bundleA.moduleHashes, bundleB.moduleHashes);
 
       if (intersection.length) {
-        let moduleHashesA = _.difference(bundleA.moduleHashes, intersection);
-        let moduleHashesB = _.difference(bundleB.moduleHashes, intersection);
+        const moduleHashesA = _.difference(bundleA.moduleHashes, intersection);
+        const moduleHashesB = _.difference(bundleB.moduleHashes, intersection);
         bundles[a] = Object.assign({}, bundleA, { moduleHashes: moduleHashesA });
         bundles[b] = Object.assign({}, bundleB, { moduleHashes: moduleHashesB });
 
@@ -28,7 +28,7 @@ function* genBundlesWithImplicit (bundles, implicitBundleDest) {
     }
   }
 
-  for (let bundle of bundles) {
+  for (const bundle of bundles) {
     if (bundle.moduleHashes.length) {
       yield bundle;
     }
