@@ -16,7 +16,7 @@ const localStorageSaveTmpl = getTemplate("local-storage-save");
 
 export default {
   localStorage: {
-    get: opts =>localStorageGetTmpl(opts || {}).body[0].expression.body.body,
+    get: opts => localStorageGetTmpl(opts || {}).body[0].expression.body.body,
     save: opts => localStorageSaveTmpl(opts || {}).body
   }
 };
