@@ -48,7 +48,7 @@ export function constructCommonModule (moduleBody, deps) {
  * @param  {Array}  commonModules  Array of object expressions.
  * @param  {String} globalName     Global variable name of the Interlock run-time.
  *
- * @return {Array[ASTnode]}        Array of AST nodes to be emitted as JavaScript.
+ * @return {Array}                 Array of AST nodes to be emitted as JavaScript.
  */
 export function constructModuleSet (commonModules, globalName) {
   // Construct an object whose keys are the module hashes and values are AST expression
@@ -59,15 +59,16 @@ export function constructModuleSet (commonModules, globalName) {
     })
   );
 
-  return moduleSetTmpl({ globalName: b.literal(globalName), modules: modules }).body;
+  return moduleSetTmpl({ globalName: b.literal(globalName), modules }).body;
 }
 
 /**
  * Construct the guts of the Interlock run-time for inclusion in file output.
  *
  * @param  {String} globalName  Global variable name of Interlock run-time.
+ * @param  {String} cacheMode   (DEPRECATED) cache mode enum.
  *
- * @return {Array[ASTnode]}
+ * @return {Array}              Array of AST nodes.
  */
 export function constructRuntime (globalName, cacheMode) {
   const cacheRuntime = cache[cacheMode];
@@ -76,8 +77,8 @@ export function constructRuntime (globalName, cacheMode) {
 
   return runtimeTmpl({
     globalName: b.literal(globalName),
-    cacheGet: cacheGet,
-    cacheSave: cacheSave
+    cacheGet,
+    cacheSave
   }).body;
 }
 
@@ -87,7 +88,7 @@ export function constructRuntime (globalName, cacheMode) {
  * @param  {String} moduleHash  Hash of the module to be required.
  * @param  {String} globalName  Global variable name of Interlock run-time.
  *
- * @return {Array[ASTnode]}
+ * @return {ASTnode}            Single AST node.
  */
 export function constructRequireStmt (moduleHash, globalName) {
   return loadModuleWhenReadyTmpl({
@@ -99,9 +100,10 @@ export function constructRequireStmt (moduleHash, globalName) {
 /**
  * Transforms a map of module-hashes-to-URLs to the AST equivalent.
  *
- * @param  {Object}  urls  Keys are module hashes, values are URL strings.
+ * @param  {Object}  urls       Keys are module hashes, values are URL strings.
+ * @param  {String} globalName  Global variable name of Interlock run-time.
  *
- * @return {ASTnode}
+ * @return {ASTnode}            Single AST node.
  */
 export function constructUrlRegistration (urls, globalName) {
   return registerUrlsTmpl({
@@ -118,7 +120,7 @@ export function constructUrlRegistration (urls, globalName) {
  * The primary constructor.  Given a set of options, construct Program AST
  * to be emitted as JavaScript.
  *
- * @param  {Object}  opts
+ * @param  {Object}  opts                 Options.
  * @param  {Boolean} opts.includeRuntime  Indicates whether Interlock run-time should be emitted.
  * @param  {Object}  opts.urls            Optional. If included, map of module hashes to URLs
  *                                        will be emitted.
@@ -126,6 +128,8 @@ export function constructUrlRegistration (urls, globalName) {
  *                                        transformed into output module AST and emitted.
  * @param  {String}  opts.initialRequire  Optional. If included, the statement in the AST output
  *                                        will be to require the specified module.
+ *
+ * @return {ASTnode}                      Single program AST node.
  */
 export function constructBundle (opts) {
   const globalName = opts.globalName || "__interlock__";
@@ -142,7 +146,7 @@ export function constructBundle (opts) {
   if (opts.modules) {
     const commonModules = _.chain(opts.modules)
       .map(function (module) {
-        var moduleAst = constructCommonModule(module.ast.body, module.dependencies);
+        const moduleAst = constructCommonModule(module.ast.body, module.dependencies);
         return [module.hash, moduleAst];
       })
       .object()
