@@ -44,7 +44,7 @@ export const getBundles = Pluggable.promise(function getBundles (bootstrappedBun
     // All subsequent methods in the chain fire 2x what they should. Identify the reason why
     // this is happening.
     .flatMap(resolved => {
-      let [{byHash}, explicitBundles] = resolved;
+      const [{byHash}, explicitBundles] = resolved;
       return this.dedupeImplicit(explicitBundles)
         // Populate bundles with module objects.
         .map(bundle => Object.assign({}, bundle, {
