@@ -6,7 +6,7 @@ export default function coerceToCommonJs (origAst) {
 
   const getRequires = new Plugin("get-requires", {
     visitor: {
-      CallExpression: function (node/*, parent */) {
+      CallExpression (node/*, parent */) {
         if (node.callee.name === "require") {
           if (node.arguments.length === 0) {
             throw new Error("Require expressions must include a target.");
