@@ -24,7 +24,7 @@ const parseSourceToAst = Pluggable.promise(function parseSourceToAst (raw, sourc
     const traversable = getTraversableAst(babelAst);
 
     return transform(traversable, tuple => {
-      const { node/*, type*/, key/*, children, parents*/ } = tuple;
+      const { node/*, type, key, children, parents*/ } = tuple;
       if (node.loc) {
         return deepAssign(tuple, "node.loc.source", sourceFile);
       }
