@@ -59,12 +59,12 @@ export default function () {
 
   return new Plugin("convert-from-amd", {
     visitor: {
-      ExpressionStatement: function (node, parent) {
+      ExpressionStatement (node, parent) {
         if (parent.type === "Program") {
           topLevelNode = node;
         }
       },
-      CallExpression: function (node, parent) {
+      CallExpression (node, parent) {
         if (parent === topLevelNode && node.callee.name === "define") {
           const args = node.arguments;
           if (args.length === 2 &&
