@@ -29,7 +29,7 @@ function normalizeEntryPoints (entryPoints) {
 /**
  * The entry point for the Interlock application
  *
- * @param   {Object}    options
+ * @param   {Object}    options               Compilation options.
  *
  * @param   {Object}    options.entry         A hash where keys are input files, relative to
  *                                            srcRoot and treated as entry points, and values
@@ -55,6 +55,8 @@ function normalizeEntryPoints (entryPoints) {
  * @param   {Array}     options.plugins       An Array of interlock Plugins
  * @param   {Boolean}   options.includeComments     Include comments in the compiled bundles
  * @param   {String}    options.implicitBundleDest  The location to emit shared dependency bundles
+ *
+ * @returns {void}
  */
 export default function Interlock (options) {
   Joi.validate(options || {}, interlockConstructorInput, (err, value) => {
@@ -94,7 +96,7 @@ Interlock.prototype.build = function () {
 };
 
 Interlock.prototype._saveBundles = function (compilation) {
-  for (let [bundleDest, bundle] of entries(compilation.bundles)) {
+  for (const [bundleDest, bundle] of entries(compilation.bundles)) {
     const bundleOutput = bundle.raw;
     const destRoot = path.join(compilation.opts.destRoot, bundleDest);
     mkdirp(path.dirname(destRoot));
@@ -105,7 +107,7 @@ Interlock.prototype._saveBundles = function (compilation) {
 
 function getRefreshedAsset (compilation, changedFilePath) {
   const origAsset = compilation.cache.modulesByAbsPath[changedFilePath];
-  let newAsset = Object.assign({}, origAsset, {
+  const newAsset = Object.assign({}, origAsset, {
     rawSource: null,
     ast: null,
     requireNodes: null,
@@ -116,7 +118,7 @@ function getRefreshedAsset (compilation, changedFilePath) {
   return loadAst.call(compilation, newAsset);
 }
 
-Interlock.prototype.watch = function (save=false) {
+Interlock.prototype.watch = function (save = false) {
   const self = this;
   let lastCompilation = null;
   const absPathToModuleHash = Object.create(null);
@@ -126,8 +128,8 @@ Interlock.prototype.watch = function (save=false) {
   return most.create(add => {
     function onCompileComplete (compilation) {
       lastCompilation = compilation;
-      for (let [, bundleObj] of entries(compilation.bundles)) {
-        for (let module of bundleObj.modules || []) {
+      for (const [, bundleObj] of entries(compilation.bundles)) {
+        for (const module of bundleObj.modules || []) {
           watcher.add(module.path);
           absPathToModuleHash[module.path] = module.hash;
         }
@@ -138,7 +140,7 @@ Interlock.prototype.watch = function (save=false) {
     }
 
     watcher.on("change", changedFilePath => {
-      for (let modulePath of Object.keys(absPathToModuleHash)) { watcher.unwatch(modulePath); }
+      for (const modulePath of Object.keys(absPathToModuleHash)) { watcher.unwatch(modulePath); }
 
       const refreshedAsset = getRefreshedAsset(lastCompilation, changedFilePath);
       delete lastCompilation.cache.modulesByAbsPath[changedFilePath];
