@@ -4,8 +4,8 @@ export const CONTINUE = Symbol.for("interlock.continue");
 function getContext (baseCxt, dependencies) {
   const context = Object.create(null);
   Object.assign(context, baseCxt);
-  for (let depKey of Object.keys(dependencies)) {
-    let dep = dependencies[depKey];
+  for (const depKey of Object.keys(dependencies)) {
+    const dep = dependencies[depKey];
     context[depKey] = dep.__isPluggable__ ? dep.bind(baseCxt) : dep;
   }
   return context;
@@ -21,7 +21,7 @@ function getPluggableSequences (invokedCxt, fnName) {
 
 export function stream (fn, dependencies = {}) {
   function pluggableFn () {
-    var invokedCxt = this || {};  // eslint-disable-line consistent-this
+    const invokedCxt = this || {};  // eslint-disable-line consistent-this
 
     const context = getContext(this, dependencies);
     const args = Array.prototype.slice.call(arguments);
@@ -30,7 +30,7 @@ export function stream (fn, dependencies = {}) {
     let chainedResult = CONTINUE;
 
     // Apply overrides, followed by default function.
-    for (let fnCandidate of overrideSeq.concat(fn)) {
+    for (const fnCandidate of overrideSeq.concat(fn)) {
       chainedResult = chainedResult === CONTINUE ?
         fnCandidate.apply(context, args) :
         chainedResult;
@@ -38,7 +38,7 @@ export function stream (fn, dependencies = {}) {
     chainedResult = chainedResult === CONTINUE ? null : chainedResult;
 
     // Apply transforms.
-    for (let transform of transformSeq) {
+    for (const transform of transformSeq) {
       chainedResult = transform.call(context, chainedResult, args);
     }
 
@@ -51,7 +51,7 @@ export function stream (fn, dependencies = {}) {
 
 export function promise (fn, dependencies = {}) {
   function pluggableFn () {
-    var invokedCxt = this || {};  // eslint-disable-line consistent-this
+    const invokedCxt = this || {};  // eslint-disable-line consistent-this
 
     const context = getContext(this, dependencies);
     const args = Array.prototype.slice.call(arguments);
@@ -60,8 +60,8 @@ export function promise (fn, dependencies = {}) {
     let chainedResult = Promise.resolve(CONTINUE);
 
     // Apply overrides, followed by default function.
-    for (let fnCandidate of overrideSeq.concat(fn)) {
-      chainedResult = chainedResult.then(previousResult => {
+    for (const fnCandidate of overrideSeq.concat(fn)) {
+      chainedResult = chainedResult.then(previousResult => { // eslint-disable-line no-loop-func
         return previousResult === CONTINUE ?
           Promise.resolve(fnCandidate.apply(context, args)) :
           previousResult;
@@ -70,8 +70,8 @@ export function promise (fn, dependencies = {}) {
     }
 
     // Apply transforms.
-    for (let transform of transformSeq) {
-      chainedResult = chainedResult.then(previousResult =>
+    for (const transform of transformSeq) {
+      chainedResult = chainedResult.then(previousResult => // eslint-disable-line no-loop-func
         Promise.resolve(transform.call(context, previousResult, args)));
     }
 
