@@ -7,8 +7,8 @@ const WINDOWS = process.platform === "win32";
 
 
 function select (items, fn) {
-  for (let item of items) {
-    let result = fn(item);
+  for (const item of items) {
+    const result = fn(item);
     if (result) { return result; }
   }
   return null;
@@ -16,13 +16,13 @@ function select (items, fn) {
 
 function isFile (absPath) {
   let stat;
-  try { stat = fs.statSync(absPath); } catch (e) { return; }
+  try { stat = fs.statSync(absPath); } catch (e) { return false; }
   return stat.isFile() || stat.isFIFO();
 }
 
 function isDir (absPath) {
   let stat;
-  try { stat = fs.statSync(absPath); } catch (e) { return; }
+  try { stat = fs.statSync(absPath); } catch (e) { return false; }
   return stat.isDirectory();
 }
 
@@ -62,8 +62,8 @@ function getNodeModulesPaths (contextPath) {
   return _.chain(0)
     .range(nestedDirs.length + 1)
     .map(idx => {
-      var pathPieces = nestedDirs.slice(0, nestedDirs.length - idx);
-      return path.resolve.apply(path, [prefix].concat(pathPieces, "node_modules"));
+      const pathPieces = nestedDirs.slice(0, nestedDirs.length - idx);
+      return path.resolve(prefix, ...pathPieces, "node_modules");
     })
     .value();
 }
@@ -74,7 +74,7 @@ function resolveSimple (requireStr, contextPath, nsRoot, extensions) {
   const resolvedPath = resolveFile(absPath, extensions) || resolveDir(absPath, extensions);
   if (resolvedPath) {
     return {
-      resolvedPath: resolvedPath,
+      resolvedPath,
       nsPath: path.relative(nsRoot, resolvedPath)
     };
   }
@@ -100,15 +100,15 @@ function resolveSimple (requireStr, contextPath, nsRoot, extensions) {
  *
  * @returns {null}               Will return null if resolve fails.
  */
-export default function resolve (requireStr, contextPath, ns, nsRoot, extensions, searchPaths=[]) {
+export default function resolve (requireStr, contextPath, ns, nsRoot, extensions, searchPaths = []) { // eslint-disable-line max-len
   const resolvedSimple = resolveSimple(requireStr, contextPath, nsRoot, extensions);
 
   if (resolvedSimple) {
     return {
       path: resolvedSimple.resolvedPath,
-      ns: ns,
+      ns,
       nsPath: resolvedSimple.nsPath,
-      nsRoot: nsRoot
+      nsRoot
     };
   }
 
@@ -124,9 +124,9 @@ export default function resolve (requireStr, contextPath, ns, nsRoot, extensions
   if (resolvedPath) {
     return {
       path: resolvedPath,
-      ns: ns,
+      ns,
       nsPath: path.relative(nsRoot, resolvedPath),
-      nsRoot: nsRoot
+      nsRoot
     };
   }
 
