@@ -1,12 +1,12 @@
 import _ from "lodash";
 
 export function* entries (obj) {
-  for (let key of Object.keys(obj)) {
+  for (const key of Object.keys(obj)) {
     yield [key, obj[key]];
   }
 }
 
-function _deepAssign(obj, keyPath, newVal) {
+function _deepAssign (obj, keyPath, newVal) {
   const key = keyPath[0];
   const modified = Object.assign({}, obj);
   if (keyPath.length === 1) {
